Move PDE workspace headers to src/pde/internal/
Restricts visibility of PDEWorkspace, AmericanPDEWorkspace,
PDESolver, SpatialOperator, and create_spatial_operator to
//src/option, //tests/internal, and //benchmarks. The public
//src/option:american_option target now depends on
//src/pde/internal:workspace via implementation_deps, preventing
transitive header exposure to library consumers.

Atomic move: all 5 headers, every #include, and the BUILD changes
land in a single commit so intermediate states build cleanly.

Also updates public headers (american_option.hpp,
american_option_batch.hpp) to remove internal PDE includes,
replacing pde_solver.hpp with the stable trbdf2_config.hpp.
