+++ src/option/BUILD.bazel
@@ -36,13 +36,10 @@ cc_library(
     hdrs = ["american_option.hpp"],
     deps = [
         "//src/support:parallel",
-        "//src/pde/core:pde_solver",
         "//src/pde/core:boundary_conditions",
-        "//src/pde/core:pde_workspace",
-        "//src/pde/core:american_pde_workspace",
         "//src/pde/core:grid",
         "//src/pde/core:time_domain",
-        "//src/pde/operators:operator_factory",
+        "//src/pde/core:trbdf2_config",
         "//src/pde/operators:black_scholes_pde",
         "//src/math:cubic_spline_solver",
         ":american_option_result",
@@ -50,6 +47,9 @@ cc_library(
         ":grid_spec_types",
         ":option_spec",
         "//src/support:error_types",
+    ],
+    implementation_deps = [
+        "//src/pde/internal:workspace",
     ],
     copts = [
         "-Wall",
@@ -91,8 +91,6 @@ cc_library(
     hdrs = ["american_option_batch.hpp"],
     deps = [
         ":american_option",
-        "//src/pde/core:pde_workspace",
-        "//src/pde/core:american_pde_workspace",
         "//src/pde/core:grid",
         "//src/support:ivcalc_trace_hdr",
         "//src/support:thread_workspace",
@@ -253,7 +251,6 @@ cc_library(
         ":iv_result",
         "//src/math:root_finding",
         ":american_option",
-        "//src/pde/core:pde_workspace",
         "//src/pde/core:grid",
         "//src/support:parallel",
         "//src/support:ivcalc_trace_hdr",
