+++ src/pde/core/BUILD.bazel
@@ -64,55 +64,3 @@ cc_library(
     include_prefix = "mango/pde/core",
 )
 
-cc_library(
-    name = "pde_workspace",
-    hdrs = ["pde_workspace.hpp"],
-    deps = ["//src/math:tridiagonal_matrix_view"],
-    copts = [
-        "-Wall",
-        "-Wextra",
-    ],
-    visibility = ["//visibility:public"],
-    strip_include_prefix = "/src/pde/core",
-    include_prefix = "mango/pde/core",
-)
-
-cc_library(
-    name = "american_pde_workspace",
-    hdrs = ["american_pde_workspace.hpp"],
-    deps = [
-        ":pde_workspace",
-        "//src/support:lifetime",
-    ],
-    copts = [
-        "-Wall",
-        "-Wextra",
-    ],
-    visibility = ["//visibility:public"],
-    strip_include_prefix = "/src/pde/core",
-    include_prefix = "mango/pde/core",
-)
-
-cc_library(
-    name = "pde_solver",
-    hdrs = ["pde_solver.hpp"],
-    deps = [
-        ":grid",
-        ":pde_workspace",
-        "//src/pde/operators:centered_difference",
-        ":boundary_conditions",
-        ":time_domain",
-        ":trbdf2_config",
-        "//src/math:thomas_solver",
-        "//src/math:root_finding",
-        "//src/math:tridiagonal_matrix_view",
-        "//src/support:error_types",
-    ],
-    copts = [
-        "-Wall",
-        "-Wextra",
-    ],
-    visibility = ["//visibility:public"],
-    strip_include_prefix = "/src/pde/core",
-    include_prefix = "mango/pde/core",
-)
