+++ src/pde/internal/american_pde_workspace.hpp
@@ -2,7 +2,7 @@
 // src/pde/core/american_pde_workspace.hpp
 #pragma once
 
-#include "mango/pde/core/pde_workspace.hpp"
+#include "mango/pde/internal/pde_workspace.hpp"
 #include "mango/support/lifetime.hpp"
 #include <span>
 #include <expected>
