+++ src/pde/internal/pde_solver.hpp
@@ -2,7 +2,7 @@
 #pragma once
 
 #include "mango/pde/core/grid.hpp"
-#include "mango/pde/core/pde_workspace.hpp"
+#include "mango/pde/internal/pde_workspace.hpp"
 #include "mango/pde/core/boundary_conditions.hpp"
 #include "mango/pde/core/time_domain.hpp"
 #include "mango/pde/core/trbdf2_config.hpp"
