+++ tests/BUILD.bazel
@@ -1234,7 +1234,6 @@ cc_test(
         "//src/option/table/bspline:bspline_surface",
         "//src/option/table/bspline:bspline_tensor_accessor",
         "//src/option:american_option",
-        "//src/pde/core:pde_workspace",
         "@googletest//:gtest",
         "@googletest//:gtest_main",
     ],
@@ -1320,7 +1319,6 @@ cc_test(
     deps = [
         "//src/option:american_option",
         "//src/option:american_option_batch",
-        "//src/pde/core:pde_workspace",
         "@googletest//:gtest",
         "@googletest//:gtest_main",
     ],
