+++ tests/internal/american_pde_workspace_test.cc
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT
 // tests/american_pde_workspace_test.cc
 
-#include "mango/pde/core/american_pde_workspace.hpp"
+#include "mango/pde/internal/american_pde_workspace.hpp"
 #include "mango/support/thread_workspace.hpp"
 #include <gtest/gtest.h>
 #include <vector>
