+++ tests/internal/pde_workspace_test.cc
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-#include "mango/pde/core/pde_workspace.hpp"
+#include "mango/pde/internal/pde_workspace.hpp"
 #include <gtest/gtest.h>
 #include <vector>
