+++ tests/internal/temporal_event_test.cc
@@ -1,13 +1,12 @@
 // SPDX-License-Identifier: MIT
 #include <gtest/gtest.h>
-#include "mango/pde/core/pde_solver.hpp"
-#include "mango/pde/core/pde_workspace.hpp"
-#include "mango/pde/core/pde_workspace.hpp"
+#include "mango/pde/internal/pde_solver.hpp"
+#include "mango/pde/internal/pde_workspace.hpp"
 #include "mango/pde/core/grid.hpp"
 #include "mango/pde/core/boundary_conditions.hpp"
 #include "mango/pde/operators/laplacian_pde.hpp"
-#include "mango/pde/operators/spatial_operator.hpp"
-#include "mango/pde/operators/operator_factory.hpp"
+#include "mango/pde/internal/spatial_operator.hpp"
+#include "mango/pde/internal/operator_factory.hpp"
 #include "mango/pde/core/grid.hpp"
 #include "mango/pde/core/time_domain.hpp"
 #include <cmath>
